+++ lib/cli/Cargo.toml
@@ -33,7 +33,7 @@ wasmer-compiler-llvm = { version = "=2.3.0", path = "../compiler-llvm", optional
 wasmer-emscripten = { version = "=2.3.0", path = "../emscripten", optional = true }
 wasmer-vm = { version = "=2.3.0", path = "../vm" }
 wasmer-wasi = { version = "=2.3.0", path = "../wasi", optional = true }
-wasmer-wasi-experimental-io-devices = { version = "=2.3.0", path = "../wasi-experimental-io-devices", optional = true }
+wasmer-wasi-experimental-io-devices = { version = "=2.3.0", path = "../wasi-experimental-io-devices", optional = true, features = ["link_external_libs"] }
 wasmer-wast = { version = "=2.3.0", path = "../../tests/lib/wast", optional = true }
 wasmer-cache = { version = "=2.3.0", path = "../cache", optional = true }
 wasmer-types = { version = "=2.3.0", path = "../types" }
